refactor: organize user/profile dir
