+++ src/app/comment/service.rs
@@ -2,9 +2,7 @@ use super::model::{Comment, CreateComment, DeleteComment};
 use crate::app::article::model::{Article, FetchBySlugAndAuthorId};
 use crate::app::user::model::User;
 use crate::appv2::features::profile::entities::Profile;
-use crate::appv2::features::profile::usecases::services::{
-    conver_user_to_profile, ConverUserToProfile,
-};
+use crate::appv2::features::profile::services::{conver_user_to_profile, ConverUserToProfile};
 use crate::error::AppError;
 use diesel::pg::PgConnection;
 use uuid::Uuid;
