+++ src/appv2/features/profile/usecases.rs
@@ -1,7 +1,7 @@
 use super::presenters::ProfilePresenter;
 use super::repositories::ProfileRepository;
 use crate::app::user::model::User;
-use crate::appv2::features::user::domains::user_repository::UserRepository;
+use crate::appv2::features::user::repositories::UserRepository;
 use crate::error::AppError;
 use actix_web::HttpResponse;
