--- src/appv2/features/user/domains/mod.rs
@@ -1,1 +1,0 @@
-pub mod user_repository;
