+++ src/appv2/features/user/mod.rs
@@ -1,1 +1,1 @@
-pub mod domains;
+pub mod repositories;
